[STORM-3595] Refactor common code into BaseResourceAwareStrategy. Overridden sortObjectResources() pulled into interface drags internal objects too. Need further refactoring.
