+++ storm-server/src/main/java/org/apache/storm/scheduler/resource/strategies/scheduling/IStrategy.java
@@ -12,10 +12,17 @@
 
 package org.apache.storm.scheduler.resource.strategies.scheduling;
 
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
+import java.util.TreeSet;
+
 import org.apache.storm.scheduler.Cluster;
+import org.apache.storm.scheduler.ExecutorDetails;
 import org.apache.storm.scheduler.TopologyDetails;
 import org.apache.storm.scheduler.resource.SchedulingResult;
+import org.apache.storm.scheduler.resource.normalization.NormalizedResourceOffer;
 
 /**
  * An interface to for implementing different scheduling strategies for the resource aware scheduling.
@@ -27,6 +34,13 @@ public interface IStrategy {
      * @param config the cluster configuration
      */
     void prepare(Map<String, Object> config);
+
+    /**
+     * Initialize for the default implementation of schedule().
+     *
+     * @param sortNodesForEachExecutor Call sortNodes() before scheduling each executor.
+     */
+    void initForSchedule(boolean sortNodesForEachExecutor);
 
     /**
      * This method is invoked to calculate a scheduling for topology td.  Cluster will reject any changes that are
@@ -41,4 +55,99 @@ public interface IStrategy {
      *     successful.
      */
     SchedulingResult schedule(Cluster schedulingState, TopologyDetails td);
+
+    /**
+     * Scheduling uses sortAllNodes() which eventually uses a sortObjectResources()
+     * method that can be overridden to provided different behavior.
+     * <p>
+     * sortObjectResources() method needs supporting classes ObjectResources,
+     * AllResources and ExistingShceduleFunc that are also defined here.
+     * </p>
+     *
+     * @param allResources         contains all individual ObjectResources as well as cumulative stats
+     * @param exec                 executor for which the sorting is done
+     * @param topologyDetails      topologyDetails for the above executor
+     * @param existingScheduleFunc a function to get existing executors already scheduled on this object
+     * @return a sorted list of ObjectResources
+     */
+    TreeSet<ObjectResources> sortObjectResources(
+            AllResources allResources, ExecutorDetails exec, TopologyDetails topologyDetails,
+            ExistingScheduleFunc existingScheduleFunc);
+
+    /**
+     * a class to contain individual object resources as well as cumulative stats.
+     */
+    class AllResources {
+        List<ObjectResources> objectResources = new LinkedList<>();
+        final NormalizedResourceOffer availableResourcesOverall;
+        final NormalizedResourceOffer totalResourcesOverall;
+        String identifier;
+
+        public AllResources(String identifier) {
+            this.identifier = identifier;
+            this.availableResourcesOverall = new NormalizedResourceOffer();
+            this.totalResourcesOverall = new NormalizedResourceOffer();
+        }
+
+        public AllResources(AllResources other) {
+            this(null,
+                    new NormalizedResourceOffer(other.availableResourcesOverall),
+                    new NormalizedResourceOffer(other.totalResourcesOverall),
+                    other.identifier);
+            List<ObjectResources> objectResourcesList = new ArrayList<>();
+            for (ObjectResources objectResource : other.objectResources) {
+                objectResourcesList.add(new ObjectResources(objectResource));
+            }
+            this.objectResources = objectResourcesList;
+        }
+
+        public AllResources(List<ObjectResources> objectResources, NormalizedResourceOffer availableResourcesOverall,
+                            NormalizedResourceOffer totalResourcesOverall, String identifier) {
+            this.objectResources = objectResources;
+            this.availableResourcesOverall = availableResourcesOverall;
+            this.totalResourcesOverall = totalResourcesOverall;
+            this.identifier = identifier;
+        }
+    }
+
+    /**
+     * class to keep track of resources on a rack or node.
+     */
+    class ObjectResources {
+        public final String id;
+        public NormalizedResourceOffer availableResources;
+        public NormalizedResourceOffer totalResources;
+        public double effectiveResources = 0.0;
+
+        public ObjectResources(String id) {
+            this.id = id;
+            this.availableResources = new NormalizedResourceOffer();
+            this.totalResources = new NormalizedResourceOffer();
+        }
+
+        public ObjectResources(ObjectResources other) {
+            this(other.id, other.availableResources, other.totalResources, other.effectiveResources);
+        }
+
+        public ObjectResources(String id, NormalizedResourceOffer availableResources, NormalizedResourceOffer totalResources,
+                               double effectiveResources) {
+            this.id = id;
+            this.availableResources = availableResources;
+            this.totalResources = totalResources;
+            this.effectiveResources = effectiveResources;
+        }
+
+        @Override
+        public String toString() {
+            return this.id;
+        }
+    }
+
+    /**
+     * interface for calculating the number of existing executors scheduled on a object (rack or node).
+     */
+    interface ExistingScheduleFunc {
+        int getNumExistingSchedule(String objectId);
+    }
+
 }
